Integrate custom code into rawSurveyData Remove code bloat
